requirements.txt
allure_python_commons==2.13.2
behave==1.2.6
selenium==4.18.1
selenium_page_factory==2.6
Faker~=23.3.0
requests~=2.31.0